removed playground button
