+++ src/app/page.js
@@ -1,13 +1,33 @@
 'use client';
+import React from 'react';
 
-import Link from 'next/link';
+import AboutUs from '@/components/about-us/about-us';
+import CampusAmbasador from '@/components/campusAmbasador/CampusAmbasador';
+import CommunityPartnerSection from '@/components/communityPartners/CommunityPartners';
+import Gallery from '@/components/gallery/gallery';
+import Hero from '@/components/hero/hero';
+import Prizes from '@/components/prizes/prizes';
+import Questions from '@/components/questions/Questions';
+import { StatsMain } from '@/components/stats/stats-main';
+import Timeline from '@/components/timeline/timeline';
+import { Sponsors, PastSponsors } from '@/components/sponsors/sponsors';
 
-export default function Home() {
+const Playground = () => {
   return (
-    <div>
-      <Link href={'/playground'}>
-        <h3>Go to Playground</h3>
-      </Link>
+    <div className='flex flex-col items-center justify-center w-full h-full px-0 md:px-5'>
+      <Hero />
+      <AboutUs />
+      <StatsMain />
+      <Prizes />
+      <Gallery />
+      <Timeline />
+      <Sponsors />
+      <PastSponsors />
+      <CommunityPartnerSection />
+      <CampusAmbasador />
+      <Questions />
     </div>
   );
-}
+};
+
+export default Playground;
